Don't export internals of embed stream in plugin-api (for now)
