--- plugin-api/beta/client/actions/stream.js
@@ -1,1 +1,0 @@
-export {setActiveTab} from 'coral-embed-stream/src/actions/stream';
