+++ plugin-api/beta/client/utils/index.js
@@ -1,0 +1,1 @@
+export {insertCommentsSorted} from 'coral-framework/utils';
