--- plugin-api/beta/client/utils/stream.js
@@ -1,1 +1,0 @@
-export {findCommentInEmbedQuery, insertSorted} from 'coral-embed-stream/src/graphql/utils';
