+++ plugins/talk-plugin-featured-comments/client/index.js
@@ -5,7 +5,8 @@ import FeaturedButton from './components/FeaturedButton';
 import translations from './translations.json';
 import update from 'immutability-helper';
 
-import {findCommentInEmbedQuery, insertCommentSorted} from 'plugin-api/beta/client/utils/stream';
+import {findCommentInEmbedQuery} from 'coral-embed-stream/src/graphql/utils';
+import {insertCommentsSorted} from 'plugin-api/beta/client/utils';
 
 export default {
   translations,
@@ -51,7 +52,7 @@ export default {
             asset: {
               featuredComments: {
                 nodes: {
-                  $apply: (nodes) => insertCommentSorted(nodes, comment, 'REVERSE_CHRONOLOGICAL')
+                  $apply: (nodes) => insertCommentsSorted(nodes, comment, 'REVERSE_CHRONOLOGICAL')
                 }
               },
               featuredCommentsCount: {
